User role authentication and access to the resource part 1

+++ Projekt Web API/Papu/Papu/Services/ProductService.cs
@@ -1,11 +1,14 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using Papu.Authorization;
 using Papu.Entities;
 using Papu.Exceptions;
 using Papu.Models;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace Papu.Services
 {
@@ -14,12 +17,16 @@ namespace Papu.Services
         private readonly PapuDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<ProductService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ProductService(PapuDbContext dbContext, IMapper mapper, ILogger<ProductService> logger)
+        //Dzięki IAuthorization (serwisowi) Asp.Net Core na podstawie wymagania
+        //będzie wstanie wywołać odpowiedni handler
+        public ProductService(PapuDbContext dbContext, IMapper mapper, ILogger<ProductService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
         //Pobranie jednego produktu po id 
@@ -59,9 +66,11 @@ namespace Papu.Services
         }
 
         //Utworzenie jednego produktu na podstawie obiektu dto 
-        public int CreateProduct(CreateProductDto dto)
+        public int CreateProduct(CreateProductDto dto, int userId)
         {
             var product = _mapper.Map<Product>(dto);
+            //Dostaniemy informację jaki użytkownik stworzył konkretny produkt w bazie danych
+            product.CreatedById = userId;
 
             Category category = _dbContext.Categories
                 .FirstOrDefault(c => c.CategoryName == dto.CategoryName);
@@ -93,8 +102,8 @@ namespace Papu.Services
         }
 
         //Edycja jednego produktu na podstawie id i obiektu dto
-        //tylko ten kto utworzył dany zasób, będzie mógł go modyfikować lub usuwać
-        public void UpdateProduct(int id, UpdateProductDto dto)
+        //Tylko ten kto utworzył dany zasób, będzie mógł go modyfikować lub usuwać
+        public void UpdateProduct(int id, UpdateProductDto dto, ClaimsPrincipal user)
         {
             var product =
                 _dbContext.Products
@@ -107,6 +116,16 @@ namespace Papu.Services
             if (product is null)
             {
                 throw new NotFoundException("Product not found");
+            }
+
+            //Sprawdzamy czy to użytkownik który stworzył dany produkt chce go zmodyfikować
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, 
+                product, new ResourceOperationRequirementProduct(ResourceOperation.Update)).Result;
+
+            //Sprawdzamy czy autoryzacja użytkownika nie powiodła się
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("This product is not your");
             }
 
             Category category = _dbContext.Categories
@@ -149,7 +168,7 @@ namespace Papu.Services
 
         //Usunięcie jednego produktu na podstawie id
         //tylko ten kto utworzył dany zasób, będzie mógł go modyfikować lub usuwać
-        public void DeleteProduct(int id)
+        public void DeleteProduct(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Product with id: {id} DELETE action invoked");
 
@@ -163,6 +182,16 @@ namespace Papu.Services
             if (product is null)
             {
                 throw new NotFoundException("Product not found");
+            }
+
+            //Sprawdzamy czy to użytkownik który stworzył dany produkt chce go usunąć
+            var authorizationResult = _authorizationService.AuthorizeAsync(user,
+                product, new ResourceOperationRequirementProduct(ResourceOperation.Delete)).Result;
+
+            //Sprawdzamy czy autoryzacja użytkownika nie powiodła się
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("This product is not your");
             }
 
             _dbContext.Products.Remove(product);
